backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.3.0
pydantic-settings==2.0.3
apscheduler==3.10.4
sqlalchemy==2.0.21
alembic==1.12.0
pymysql==1.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
email-validator==2.0.0
pytest==7.4.2
httpx==0.25.0
